refactor(config): move unknown-key recognition out of the didChange handler
Structural only; no behavior change (3117 lib tests, unchanged count).

The known-key sets and the walker over them described a property of the
configuration schema, not of `workspace/didChangeConfiguration`, and the
explicit `--config-file` path needs the same answer. They now live in
`config::unknown_keys`, together with the schema-consistency tests that
assert each set still matches its struct.
